requirements.txt
# AI APIs
google-generativeai>=0.8.0
openai>=1.40.0
# Fish Audio TTS (httpx for API calls)
ormsgpack>=1.5.0

# Image Processing
pillow>=10.4.0

# Video Processing (FFmpeg wrapper)
ffmpeg-python>=0.2.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Data Validation
pydantic>=2.8.0

# HTTP Requests
requests>=2.32.0
httpx>=0.27.0

# Async Support
aiohttp>=3.10.0
aiofiles>=24.1.0

# Utilities
tqdm>=4.66.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Type Checking
mypy>=1.11.0
